+++ src/ActivityLogger.php
@@ -39,7 +39,7 @@ class ActivityLogger
 
         $this->properties = collect();
 
-        $this->authDriver = $config['laravel-activitylog']['default_auth_driver'] ?? $auth->getDefaultDriver();
+        $this->authDriver = $config['activitylog']['default_auth_driver'] ?? $auth->getDefaultDriver();
 
         if (starts_with(app()->version(), '5.1')) {
             $this->causedBy = $auth->driver($this->authDriver)->user();
@@ -47,9 +47,9 @@ class ActivityLogger
             $this->causedBy = $auth->guard($this->authDriver)->user();
         }
 
-        $this->logName = $config['laravel-activitylog']['default_log_name'];
+        $this->logName = $config['activitylog']['default_log_name'];
 
-        $this->logEnabled = $config['laravel-activitylog']['enabled'] ?? true;
+        $this->logEnabled = $config['activitylog']['enabled'] ?? true;
     }
 
     public function performedOn(Model $model)