+++ src/CleanActivitylogCommand.php
@@ -25,7 +25,7 @@ class CleanActivitylogCommand extends Command
     {
         $this->comment('Cleaning activity log...');
 
-        $maxAgeInDays = config('laravel-activitylog.delete_records_older_than_days');
+        $maxAgeInDays = config('activitylog.delete_records_older_than_days');
 
         $cutOffDate = Carbon::now()->subDays($maxAgeInDays)->format('Y-m-d H:i:s');