+++ tests/CleanActivitylogCommandTest.php
@@ -14,7 +14,7 @@ class CleanActivitylogCommandTest extends TestCase
 
         Carbon::setTestNow(Carbon::create(2016, 1, 1, 00, 00, 00));
 
-        $this->app['config']->set('laravel-activitylog.delete_records_older_than_days', 31);
+        $this->app['config']->set('activitylog.delete_records_older_than_days', 31);
     }
 
     /** @test */